+++ apps/cudpp_testrig/test_radixsort.cpp
@@ -18,7 +18,7 @@
 
 template <typename T>
 int radixSortTest(CUDPPHandle plan, CUDPPConfiguration config, size_t *tests, 
-                  unsigned int numTests, unsigned int numElements, unsigned int keybits,
+                  unsigned int numTests, size_t numElements, unsigned int keybits,
                   testrigOptions testOptions, bool quiet)
 {
     int retval = 0;
